Refactor and fix
